Add EvaluatedRef to clean Evaluated binary operation codes
